door_backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
channels[daphne]>=4.0
channels-redis>=4.1
celery>=5.3
redis>=5.0
psycopg2-binary>=2.9
django-celery-beat>=2.5
django-celery-results>=2.5
django-cors-headers>=4.3
django-filter>=23.3
django-storages>=1.14
Pillow>=10.0
qrcode[pil]>=7.4
cryptography>=41.0
firebase-admin>=6.2
whitenoise>=6.6
django-environ>=0.11
drf-spectacular>=0.27
drf-nested-routers>=0.94
celery-progress>=0.4
django-rosetta>=0.9
